refactor(frontend): apply Vercel React best practices for bundle and render optimization
Lazy-load LicenseDisclaimer modal via React.lazy to split it into a
separate chunk, remove dead code from map-style.ts, eliminate unnecessary
global state dispatch that triggered redundant re-renders, memoize
sortedYears and keyEvents sorting, remove superfluous Fragment wrapper,
and replace barrel file with direct type imports.
