+++ apps/frontend/src/components/territory-info/territory-info-panel.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import { describe, expect, it, vi } from 'vitest';
-import type { TerritoryDescription } from '@/types';
+import type { TerritoryDescription } from '@/types/territory';
 
 // Mock the app state context
 const mockSetInfoPanelOpen = vi.fn();
