+++ apps/frontend/src/components/year-selector/year-selector.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import { fn } from 'storybook/test';
 import { AppStateProvider } from '../../contexts/app-state-context';
-import type { YearEntry } from '../../types';
+import type { YearEntry } from '../../types/year';
 import { YearSelector } from './year-selector';
 
 const sampleYears: YearEntry[] = [
