--- apps/frontend/src/types/index.ts
@@ -1,13 +1,0 @@
-export type {
-  AppState,
-  AppStateActions,
-  MapViewState,
-} from './app-state.ts';
-export { initialAppState } from './app-state.ts';
-export type {
-  KeyEvent,
-  TerritoryDescription,
-  TerritoryProperties,
-  YearDescriptionBundle,
-} from './territory.ts';
-export type { YearEntry, YearIndex } from './year.ts';
